requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
attrs==21.4.0
black==22.3.0
boto3==1.24.31
botocore==1.27.31
certifi==2022.5.18.1
chardet==3.0.4
click==8.1.3
configparser==5.0.0
execnet==1.9.0
idna==2.10
importlib-metadata==4.11.4
iniconfig==1.1.1
jmespath==1.0.1
more-itertools==8.13.0
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==6.0.2
pytest-forked==1.4.0
pytest-html==3.0.0
pytest-metadata==2.0.1
pytest-rerunfailures==9.1.1
pytest-testrail==2.9.0
pytest-xdist==2.1.0
python-dateutil==2.8.0
pytz==2022.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
simplejson==3.17.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.2.0
urllib3==1.25.11
zipp==3.8.0
